@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -28,35 +28,39 @@ dependencies:
     path: ./packages/formgen/
 
   # MobX
-  mobx: ^1.2.1
-  flutter_mobx: ^1.1.0
+  mobx: ^2.0.0-nullsafety.6
+  flutter_mobx: ^2.0.0-nullsafety.3
 
   # Tools
-  flutter_hooks: ^0.9.0
+  flutter_hooks: ^0.16.0
   freezed_annotation:
-  uuid:
-  get_it: ^4.0.2
-  intl: ^0.16.1
-  logger: ^0.9.1
-  google_fonts: 
+  uuid: ^3.0.0
+  get_it: ^6.0.0
+  intl: ^0.17.0
+  logger: ^1.0.0
+  google_fonts: ^2.0.0
   touchable:
     path: ./packages/touchable/
-  flutter_webrtc:
-    # path: ./packages/flutter-webrtc/
-    git:
-      url: https://github.com/flutter-webrtc/flutter-webrtc
-      ref: 0bc90b328e44af426083e2bbcf3f5687ec65a1c8
+  flutter_webrtc: ^0.6.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.2
+  url_strategy: ^0.2.0
 
   artemis:
-  gql_websocket_link: ^0.1.8-alpha+1612972259626
+    git:
+      url: https://github.com/glynskyi/artemis
+      ref: c769cb40a8ea0ab856da0c0e5fbdd8e368ed529b
+  gql_websocket_link:
+    git:
+      url: https://github.com/gql-dart/gql
+      ref: ce1b68b814ddf8b1056e9bc6b9a1146c2cc25d18
+      path: ./links/gql_websocket_link
   json_annotation:
   equatable:
   meta:
-  gql:
+  gql: ^0.13.0-nullsafety.2
 
 dev_dependencies:  
   flutter_test:
@@ -65,7 +69,7 @@ dev_dependencies:
   # fvm flutter pub run build_runner watch --delete-conflicting-outputs
   build_runner:
   json_serializable:
-  mobx_codegen: ^1.1.0
+  mobx_codegen: ^2.0.0-nullsafety.2
   freezed:
   lint: ^1.0.0
   integration_test:
@@ -73,9 +77,6 @@ dev_dependencies:
   formgen_generation:
     path: ./packages/formgen_generation/
 
-dependency_overrides:
-  analyzer: ">=0.41.1"
-  crypto: ">=3.0.0"
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
@@ -86,9 +87,8 @@ flutter:
   # the material Icons class.
   uses-material-design: true
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  assets:
+    - google_fonts/
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
   # For details regarding adding assets from package dependencies, see
