@@ -1,6 +1,5 @@
 import 'dart:typed_data';
 
-import 'package:meta/meta.dart';
 import 'package:neural_graph/file_system_access_chrome/file_system_write_params.dart';
 
 export 'package:neural_graph/file_system_access_chrome/file_system_write_params.dart';
@@ -24,12 +23,12 @@ abstract class FileSystemWriteChunkType {
   dynamic get value;
 
   T when<T>({
-    @required T Function(ByteBuffer value) bufferSource,
-    @required T Function(dynamic /*html.Blob*/ value) blob,
-    @required T Function(String value) string,
-    @required T Function(WriteParams value) writeParams,
+    required T Function(ByteBuffer value) bufferSource,
+    required T Function(dynamic /*html.Blob*/ value) blob,
+    required T Function(String value) string,
+    required T Function(WriteParams value) writeParams,
   }) {
-    final v = this;
+    final FileSystemWriteChunkType v = this;
     if (v is _BufferSource) return bufferSource(v.value);
     if (v is _Blob) return blob(v.value);
     if (v is _String) return string(v.value);
@@ -37,14 +36,14 @@ abstract class FileSystemWriteChunkType {
     throw "";
   }
 
-  T maybeWhen<T>({
-    T Function() orElse,
-    T Function(ByteBuffer value) bufferSource,
-    T Function(dynamic /*html.Blob*/ value) blob,
-    T Function(String value) string,
-    T Function(WriteParams value) writeParams,
+  T? maybeWhen<T>({
+    T Function()? orElse,
+    T Function(ByteBuffer value)? bufferSource,
+    T Function(dynamic /*html.Blob*/ value)? blob,
+    T Function(String value)? string,
+    T Function(WriteParams value)? writeParams,
   }) {
-    final v = this;
+    final FileSystemWriteChunkType v = this;
     if (v is _BufferSource)
       return bufferSource != null ? bufferSource(v.value) : orElse?.call();
     if (v is _Blob) return blob != null ? blob(v.value) : orElse?.call();
@@ -55,12 +54,12 @@ abstract class FileSystemWriteChunkType {
   }
 
   T map<T>({
-    @required T Function(_BufferSource value) bufferSource,
-    @required T Function(_Blob value) blob,
-    @required T Function(_String value) string,
-    @required T Function(_WriteParams value) writeParams,
+    required T Function(_BufferSource value) bufferSource,
+    required T Function(_Blob value) blob,
+    required T Function(_String value) string,
+    required T Function(_WriteParams value) writeParams,
   }) {
-    final v = this;
+    final FileSystemWriteChunkType v = this;
     if (v is _BufferSource) return bufferSource(v);
     if (v is _Blob) return blob(v);
     if (v is _String) return string(v);
@@ -68,14 +67,14 @@ abstract class FileSystemWriteChunkType {
     throw "";
   }
 
-  T maybeMap<T>({
-    T Function() orElse,
-    T Function(_BufferSource value) bufferSource,
-    T Function(_Blob value) blob,
-    T Function(_String value) string,
-    T Function(_WriteParams value) writeParams,
+  T? maybeMap<T>({
+    T Function()? orElse,
+    T Function(_BufferSource value)? bufferSource,
+    T Function(_Blob value)? blob,
+    T Function(_String value)? string,
+    T Function(_WriteParams value)? writeParams,
   }) {
-    final v = this;
+    final FileSystemWriteChunkType v = this;
     if (v is _BufferSource)
       return bufferSource != null ? bufferSource(v) : orElse?.call();
     if (v is _Blob) return blob != null ? blob(v) : orElse?.call();
