@@ -12,12 +12,12 @@ class LoopBackSample extends StatefulWidget {
 }
 
 class _MyAppState extends State<LoopBackSample> {
-  MediaStream _localStream;
-  RTCPeerConnection _peerConnection;
+  MediaStream? _localStream;
+  RTCPeerConnection? _peerConnection;
   final _localRenderer = RTCVideoRenderer();
   final _remoteRenderer = RTCVideoRenderer();
   bool _inCalling = false;
-  Timer _timer;
+  late Timer _timer;
 
   @override
   void initState() {
@@ -101,8 +101,8 @@ class _MyAppState extends State<LoopBackSample> {
       print('onCandidate: complete!');
       return;
     }
-    print('onCandidate: ' + candidate.candidate);
-    _peerConnection.addCandidate(candidate);
+    print('onCandidate: ' + candidate.candidate!);
+    _peerConnection!.addCandidate(candidate);
   }
 
   void _onTrack(RTCTrackEvent event) {
@@ -161,21 +161,21 @@ class _MyAppState extends State<LoopBackSample> {
       _peerConnection =
           await createPeerConnection(configuration, loopbackConstraints);
 
-      _peerConnection.onSignalingState = _onSignalingState;
-      _peerConnection.onIceGatheringState = _onIceGatheringState;
-      _peerConnection.onIceConnectionState = _onIceConnectionState;
-      _peerConnection.onConnectionState = _onPeerConnectionState;
-      _peerConnection.onAddStream = _onAddStream;
-      _peerConnection.onRemoveStream = _onRemoveStream;
-      _peerConnection.onIceCandidate = _onCandidate;
-      _peerConnection.onRenegotiationNeeded = _onRenegotiationNeeded;
+      _peerConnection!.onSignalingState = _onSignalingState;
+      _peerConnection!.onIceGatheringState = _onIceGatheringState;
+      _peerConnection!.onIceConnectionState = _onIceConnectionState;
+      _peerConnection!.onConnectionState = _onPeerConnectionState;
+      _peerConnection!.onAddStream = _onAddStream;
+      _peerConnection!.onRemoveStream = _onRemoveStream;
+      _peerConnection!.onIceCandidate = _onCandidate;
+      _peerConnection!.onRenegotiationNeeded = _onRenegotiationNeeded;
 
-      _peerConnection.onTrack = _onTrack;
+      _peerConnection!.onTrack = _onTrack;
 
       _localStream =
           await navigator.mediaDevices.getUserMedia(mediaConstraints);
       _localRenderer.srcObject = _localStream;
-      await _peerConnection.addStream(_localStream);
+      await _peerConnection!.addStream(_localStream!);
       /* old API
       await _peerConnection.addStream(_localStream);
       // Unified-Plan
@@ -238,13 +238,13 @@ class _MyAppState extends State<LoopBackSample> {
           init:
               RTCRtpTransceiverInit(direction: TransceiverDirection.RecvOnly));
       */
-      var description = await _peerConnection.createOffer(offerSdpConstraints);
+      var description = await _peerConnection!.createOffer(offerSdpConstraints);
       var sdp = description.sdp;
       print('sdp = $sdp');
-      await _peerConnection.setLocalDescription(description);
+      await _peerConnection!.setLocalDescription(description);
       //change for loopback.
       description.type = 'answer';
-      await _peerConnection.setRemoteDescription(description);
+      await _peerConnection!.setRemoteDescription(description);
 
       /* Unfied-Plan replaceTrack
       var stream = await MediaDevices.getDisplayMedia(mediaConstraints);
@@ -266,8 +266,8 @@ class _MyAppState extends State<LoopBackSample> {
 
   void _hangUp() async {
     try {
-      await _localStream.dispose();
-      await _peerConnection.close();
+      await _localStream!.dispose();
+      await _peerConnection!.close();
       _peerConnection = null;
       _localRenderer.srcObject = null;
       _remoteRenderer.srcObject = null;
@@ -282,7 +282,7 @@ class _MyAppState extends State<LoopBackSample> {
 
   void _sendDtmf() async {
     var dtmfSender =
-        _peerConnection.createDtmfSender(_localStream.getAudioTracks()[0]);
+        _peerConnection!.createDtmfSender(_localStream!.getAudioTracks()[0]);
     await dtmfSender.insertDTMF('123#');
   }
 
