@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.8
+
 
 import 'dart:math' as math;
 
@@ -48,7 +48,7 @@ class InteractiveViewer extends StatefulWidget {
   ///
   /// The [child] parameter must not be null.
   InteractiveViewer({
-    Key key,
+    Key? key,
     this.alignPanAxis = false,
     this.boundaryMargin = EdgeInsets.zero,
     this.constrained = true,
@@ -62,7 +62,7 @@ class InteractiveViewer extends StatefulWidget {
     this.panEnabled = true,
     this.scaleEnabled = true,
     this.transformationController,
-    @required this.child,
+    required this.child,
   })  : assert(alignPanAxis != null),
         assert(child != null),
         assert(constrained != null),
@@ -224,7 +224,7 @@ class InteractiveViewer extends StatefulWidget {
   ///
   ///  * [onInteractionStart], which handles the start of the same interaction.
   ///  * [onInteractionUpdate], which handles an update to the same interaction.
-  final GestureScaleEndCallback onInteractionEnd;
+  final GestureScaleEndCallback? onInteractionEnd;
 
   /// Called when the user begins a pan or scale gesture on the widget.
   ///
@@ -234,7 +234,7 @@ class InteractiveViewer extends StatefulWidget {
   ///
   ///  * [onInteractionUpdate], which handles an update to the same interaction.
   ///  * [onInteractionEnd], which handles the end of the same interaction.
-  final GestureScaleStartCallback onInteractionStart;
+  final GestureScaleStartCallback? onInteractionStart;
 
   /// Called when the user updates a pan or scale gesture on the widget.
   ///
@@ -244,7 +244,7 @@ class InteractiveViewer extends StatefulWidget {
   ///
   ///  * [onInteractionStart], which handles the start of the same interaction.
   ///  * [onInteractionEnd], which handles the end of the same interaction.
-  final GestureScaleUpdateCallback onInteractionUpdate;
+  final GestureScaleUpdateCallback? onInteractionUpdate;
 
   /// A [TransformationController] for the transformation performed on the
   /// child.
@@ -356,7 +356,7 @@ class InteractiveViewer extends StatefulWidget {
   ///
   ///  * [ValueNotifier], the parent class of TransformationController.
   ///  * [TextEditingController] for an example of another similar pattern.
-  final TransformationController transformationController;
+  final TransformationController? transformationController;
 
   /// Returns the closest point to the given point on the given line segment.
   @visibleForTesting
@@ -449,7 +449,7 @@ class InteractiveViewer extends StatefulWidget {
   /// Get the point inside (inclusively) the given Quad that is nearest to the
   /// given Vector3.
   @visibleForTesting
-  static Vector3 getNearestPointInside(Vector3 point, Quad quad) {
+  static Vector3? getNearestPointInside(Vector3 point, Quad quad) {
     // If the point is inside the axis aligned bounding box, then it's ok where
     // it is.
     if (pointIsInside(point, quad)) {
@@ -464,7 +464,7 @@ class InteractiveViewer extends StatefulWidget {
       InteractiveViewer.getNearestPointOnLine(point, quad.point3, quad.point0),
     ];
     double minDistance = double.infinity;
-    Vector3 closestOverall;
+    Vector3? closestOverall;
     for (final Vector3 closePoint in closestPoints) {
       final double distance = math.sqrt(
         math.pow(point.x - closePoint.x, 2) +
@@ -484,18 +484,18 @@ class InteractiveViewer extends StatefulWidget {
 
 class _InteractiveViewerState extends State<InteractiveViewer>
     with TickerProviderStateMixin {
-  TransformationController _transformationController;
+  TransformationController? _transformationController;
 
   final GlobalKey _childKey = GlobalKey();
   final GlobalKey _parentKey = GlobalKey();
-  Animation<Offset> _animation;
-  AnimationController _controller;
-  Axis _panAxis; // Used with alignPanAxis.
-  Offset _referenceFocalPoint; // Point where the current gesture began.
-  double _scaleStart; // Scale value at start of scaling gesture.
-  double _rotationStart = 0.0; // Rotation at start of rotation gesture.
+  Animation<Offset>? _animation;
+  late AnimationController _controller;
+  Axis? _panAxis; // Used with alignPanAxis.
+  Offset? _referenceFocalPoint; // Point where the current gesture began.
+  double? _scaleStart; // Scale value at start of scaling gesture.
+  double? _rotationStart = 0.0; // Rotation at start of rotation gesture.
   double _currentRotation = 0.0; // Rotation of _transformationController.value.
-  _GestureType _gestureType;
+  _GestureType? _gestureType;
 
   // TODO(justinmc): Add rotateEnabled parameter to the widget and remove this
   // hardcoded value when the rotation feature is implemented.
@@ -516,7 +516,7 @@ class _InteractiveViewerState extends State<InteractiveViewer>
     assert(!widget.boundaryMargin.bottom.isNaN);
 
     final RenderBox childRenderBox =
-        _childKey.currentContext.findRenderObject() as RenderBox;
+        _childKey.currentContext!.findRenderObject() as RenderBox;
     final Size childSize = childRenderBox.size;
     final Rect boundaryRect =
         widget.boundaryMargin.inflateRect(Offset.zero & childSize);
@@ -536,7 +536,7 @@ class _InteractiveViewerState extends State<InteractiveViewer>
   Rect get _viewport {
     assert(_parentKey.currentContext != null);
     final RenderBox parentRenderBox =
-        _parentKey.currentContext.findRenderObject() as RenderBox;
+        _parentKey.currentContext!.findRenderObject() as RenderBox;
     return Offset.zero & parentRenderBox.size;
   }
 
@@ -645,12 +645,12 @@ class _InteractiveViewerState extends State<InteractiveViewer>
     // Don't allow a scale that results in an overall scale beyond min/max
     // scale.
     final double currentScale =
-        _transformationController.value.getMaxScaleOnAxis();
+        _transformationController!.value.getMaxScaleOnAxis();
     final double totalScale = currentScale * scale;
     final double clampedTotalScale = totalScale.clamp(
       widget.minScale,
       widget.maxScale,
-    ) as double;
+    );
     final double clampedScale = clampedTotalScale / currentScale;
     final Matrix4 nextMatrix = matrix.clone()..scale(clampedScale);
 
@@ -674,7 +674,7 @@ class _InteractiveViewerState extends State<InteractiveViewer>
     if (rotation == 0) {
       return matrix.clone();
     }
-    final Offset focalPointScene = _transformationController.toScene(
+    final Offset focalPointScene = _transformationController!.toScene(
       focalPoint,
     );
     return matrix.clone()
@@ -684,7 +684,7 @@ class _InteractiveViewerState extends State<InteractiveViewer>
   }
 
   // Returns true iff the given _GestureType is enabled.
-  bool _gestureIsSupported(_GestureType gestureType) {
+  bool _gestureIsSupported(_GestureType? gestureType) {
     switch (gestureType) {
       case _GestureType.rotate:
         return _rotateEnabled;
@@ -702,7 +702,7 @@ class _InteractiveViewerState extends State<InteractiveViewer>
   // with GestureDetector's scale gesture.
   void _onScaleStart(ScaleStartDetails details) {
     if (widget.onInteractionStart != null) {
-      widget.onInteractionStart(details);
+      widget.onInteractionStart!(details);
     }
 
     if (_controller.isAnimating) {
@@ -714,8 +714,8 @@ class _InteractiveViewerState extends State<InteractiveViewer>
 
     _gestureType = null;
     _panAxis = null;
-    _scaleStart = _transformationController.value.getMaxScaleOnAxis();
-    _referenceFocalPoint = _transformationController.toScene(
+    _scaleStart = _transformationController!.value.getMaxScaleOnAxis();
+    _referenceFocalPoint = _transformationController!.toScene(
       details.localFocalPoint,
     );
     _rotationStart = _currentRotation;
@@ -724,17 +724,17 @@ class _InteractiveViewerState extends State<InteractiveViewer>
   // Handle an update to an ongoing gesture. All of pan, scale, and rotate are
   // handled with GestureDetector's scale gesture.
   void _onScaleUpdate(ScaleUpdateDetails details) {
-    final double scale = _transformationController.value.getMaxScaleOnAxis();
+    final double scale = _transformationController!.value.getMaxScaleOnAxis();
     if (widget.onInteractionUpdate != null) {
-      widget.onInteractionUpdate(ScaleUpdateDetails(
-        focalPoint: _transformationController.toScene(
+      widget.onInteractionUpdate!(ScaleUpdateDetails(
+        focalPoint: _transformationController!.toScene(
           details.localFocalPoint,
         ),
         scale: details.scale,
         rotation: details.rotation,
       ));
     }
-    final Offset focalPointScene = _transformationController.toScene(
+    final Offset focalPointScene = _transformationController!.toScene(
       details.localFocalPoint,
     );
     _gestureType ??= _getGestureType(
@@ -757,10 +757,10 @@ class _InteractiveViewerState extends State<InteractiveViewer>
         // details.scale gives us the amount to change the scale as of the
         // start of this gesture, so calculate the amount to scale as of the
         // previous call to _onScaleUpdate.
-        final double desiredScale = _scaleStart * details.scale;
+        final double desiredScale = _scaleStart! * details.scale;
         final double scaleChange = desiredScale / scale;
-        _transformationController.value = _matrixScale(
-          _transformationController.value,
+        _transformationController!.value = _matrixScale(
+          _transformationController!.value,
           scaleChange,
         );
 
@@ -768,12 +768,12 @@ class _InteractiveViewerState extends State<InteractiveViewer>
         // the same places in the scene. That means that the focal point of
         // the scale should be on the same place in the scene before and after
         // the scale.
-        final Offset focalPointSceneScaled = _transformationController.toScene(
+        final Offset focalPointSceneScaled = _transformationController!.toScene(
           details.localFocalPoint,
         );
-        _transformationController.value = _matrixTranslate(
-          _transformationController.value,
-          focalPointSceneScaled - _referenceFocalPoint,
+        _transformationController!.value = _matrixTranslate(
+          _transformationController!.value,
+          focalPointSceneScaled - _referenceFocalPoint!,
         );
 
         // details.localFocalPoint should now be at the same location as the
@@ -781,10 +781,10 @@ class _InteractiveViewerState extends State<InteractiveViewer>
         // the translate came in contact with a boundary. In that case, update
         // _referenceFocalPoint so subsequent updates happen in relation to
         // the new effective focal point.
-        final Offset focalPointSceneCheck = _transformationController.toScene(
+        final Offset focalPointSceneCheck = _transformationController!.toScene(
           details.localFocalPoint,
         );
-        if (_round(_referenceFocalPoint) != _round(focalPointSceneCheck)) {
+        if (_round(_referenceFocalPoint!) != _round(focalPointSceneCheck)) {
           _referenceFocalPoint = focalPointSceneCheck;
         }
         return;
@@ -793,9 +793,9 @@ class _InteractiveViewerState extends State<InteractiveViewer>
         if (details.rotation == 0.0) {
           return;
         }
-        final double desiredRotation = _rotationStart + details.rotation;
-        _transformationController.value = _matrixRotate(
-          _transformationController.value,
+        final double desiredRotation = _rotationStart! + details.rotation;
+        _transformationController!.value = _matrixRotate(
+          _transformationController!.value,
           _currentRotation - desiredRotation,
           details.localFocalPoint,
         );
@@ -808,12 +808,12 @@ class _InteractiveViewerState extends State<InteractiveViewer>
         }
         // Translate so that the same point in the scene is underneath the
         // focal point before and after the movement.
-        final Offset translationChange = focalPointScene - _referenceFocalPoint;
-        _transformationController.value = _matrixTranslate(
-          _transformationController.value,
+        final Offset translationChange = focalPointScene - _referenceFocalPoint!;
+        _transformationController!.value = _matrixTranslate(
+          _transformationController!.value,
           translationChange,
         );
-        _referenceFocalPoint = _transformationController.toScene(
+        _referenceFocalPoint = _transformationController!.toScene(
           details.localFocalPoint,
         );
         return;
@@ -824,7 +824,7 @@ class _InteractiveViewerState extends State<InteractiveViewer>
   // are handled with GestureDetector's scale gesture.
   void _onScaleEnd(ScaleEndDetails details) {
     if (widget.onInteractionEnd != null) {
-      widget.onInteractionEnd(details);
+      widget.onInteractionEnd!(details);
     }
     _scaleStart = null;
     _rotationStart = null;
@@ -846,7 +846,7 @@ class _InteractiveViewerState extends State<InteractiveViewer>
     }
 
     final Vector3 translationVector =
-        _transformationController.value.getTranslation();
+        _transformationController!.value.getTranslation();
     final Offset translation = Offset(translationVector.x, translationVector.y);
     final FrictionSimulation frictionSimulationX = FrictionSimulation(
       _kDrag,
@@ -870,7 +870,7 @@ class _InteractiveViewerState extends State<InteractiveViewer>
       curve: Curves.decelerate,
     ));
     _controller.duration = Duration(milliseconds: (tFinal * 1000).round());
-    _animation.addListener(_onAnimate);
+    _animation!.addListener(_onAnimate);
     _controller.forward();
   }
 
@@ -881,27 +881,27 @@ class _InteractiveViewerState extends State<InteractiveViewer>
     }
     if (event is PointerScrollEvent) {
       final RenderBox childRenderBox =
-          _childKey.currentContext.findRenderObject() as RenderBox;
+          _childKey.currentContext!.findRenderObject() as RenderBox;
       final Size childSize = childRenderBox.size;
       final double scaleChange = 1.0 - event.scrollDelta.dy / childSize.height;
       if (scaleChange == 0.0) {
         return;
       }
-      final Offset focalPointScene = _transformationController.toScene(
+      final Offset focalPointScene = _transformationController!.toScene(
         event.localPosition,
       );
-      _transformationController.value = _matrixScale(
-        _transformationController.value,
+      _transformationController!.value = _matrixScale(
+        _transformationController!.value,
         scaleChange,
       );
 
       // After scaling, translate such that the event's position is at the
       // same scene point before and after the scale.
-      final Offset focalPointSceneScaled = _transformationController.toScene(
+      final Offset focalPointSceneScaled = _transformationController!.toScene(
         event.localPosition,
       );
-      _transformationController.value = _matrixTranslate(
-        _transformationController.value,
+      _transformationController!.value = _matrixTranslate(
+        _transformationController!.value,
         focalPointSceneScaled - focalPointScene,
       );
     }
@@ -918,17 +918,17 @@ class _InteractiveViewerState extends State<InteractiveViewer>
     }
     // Translate such that the resulting translation is _animation.value.
     final Vector3 translationVector =
-        _transformationController.value.getTranslation();
+        _transformationController!.value.getTranslation();
     final Offset translation = Offset(translationVector.x, translationVector.y);
-    final Offset translationScene = _transformationController.toScene(
+    final Offset translationScene = _transformationController!.toScene(
       translation,
     );
-    final Offset animationScene = _transformationController.toScene(
-      _animation.value,
+    final Offset animationScene = _transformationController!.toScene(
+      _animation!.value,
     );
     final Offset translationChangeScene = animationScene - translationScene;
-    _transformationController.value = _matrixTranslate(
-      _transformationController.value,
+    _transformationController!.value = _matrixTranslate(
+      _transformationController!.value,
       translationChangeScene,
     );
   }
@@ -945,7 +945,7 @@ class _InteractiveViewerState extends State<InteractiveViewer>
 
     _transformationController =
         widget.transformationController ?? TransformationController();
-    _transformationController.addListener(_onTransformationControllerChange);
+    _transformationController!.addListener(_onTransformationControllerChange);
     _controller = AnimationController(
       vsync: this,
     );
@@ -958,26 +958,26 @@ class _InteractiveViewerState extends State<InteractiveViewer>
     // transformationControllers.
     if (oldWidget.transformationController == null) {
       if (widget.transformationController != null) {
-        _transformationController
+        _transformationController!
             .removeListener(_onTransformationControllerChange);
-        _transformationController.dispose();
+        _transformationController!.dispose();
         _transformationController = widget.transformationController;
-        _transformationController
+        _transformationController!
             .addListener(_onTransformationControllerChange);
       }
     } else {
       if (widget.transformationController == null) {
-        _transformationController
+        _transformationController!
             .removeListener(_onTransformationControllerChange);
         _transformationController = TransformationController();
-        _transformationController
+        _transformationController!
             .addListener(_onTransformationControllerChange);
       } else if (widget.transformationController !=
           oldWidget.transformationController) {
-        _transformationController
+        _transformationController!
             .removeListener(_onTransformationControllerChange);
         _transformationController = widget.transformationController;
-        _transformationController
+        _transformationController!
             .addListener(_onTransformationControllerChange);
       }
     }
@@ -986,9 +986,9 @@ class _InteractiveViewerState extends State<InteractiveViewer>
   @override
   void dispose() {
     _controller.dispose();
-    _transformationController.removeListener(_onTransformationControllerChange);
+    _transformationController!.removeListener(_onTransformationControllerChange);
     if (widget.transformationController == null) {
-      _transformationController.dispose();
+      _transformationController!.dispose();
     }
     super.dispose();
   }
@@ -996,7 +996,7 @@ class _InteractiveViewerState extends State<InteractiveViewer>
   @override
   Widget build(BuildContext context) {
     Widget child = Transform(
-      transform: _transformationController.value,
+      transform: _transformationController!.value,
       child: KeyedSubtree(
         key: _childKey,
         child: widget.child,
@@ -1047,7 +1047,7 @@ class TransformationController extends ValueNotifier<Matrix4> {
   ///
   /// The [value] defaults to the identity matrix, which corresponds to no
   /// transformation.
-  TransformationController([Matrix4 value])
+  TransformationController([Matrix4? value])
       : super(value ?? Matrix4.identity());
 
   /// Return the scene point at the given viewport point.
@@ -1187,7 +1187,7 @@ Offset _exceedsBy(Quad boundary, Quad viewport) {
   Offset largestExcess = Offset.zero;
   for (final Vector3 point in viewportPoints) {
     final Vector3 pointInside =
-        InteractiveViewer.getNearestPointInside(point, boundary);
+        InteractiveViewer.getNearestPointInside(point, boundary)!;
     final Offset excess = Offset(
       pointInside.x - point.x,
       pointInside.y - point.y,
@@ -1214,7 +1214,7 @@ Offset _round(Offset offset) {
 
 // Align the given offset to the given axis by allowing movement only in the
 // axis direction.
-Offset _alignAxis(Offset offset, Axis axis) {
+Offset _alignAxis(Offset offset, Axis? axis) {
   switch (axis) {
     case Axis.horizontal:
       return Offset(offset.dx, 0.0);
@@ -1226,11 +1226,11 @@ Offset _alignAxis(Offset offset, Axis axis) {
 
 // Given two points, return the axis where the distance between the points is
 // greatest. If they are equal, return null.
-Axis _getPanAxis(Offset point1, Offset point2) {
+Axis? _getPanAxis(Offset? point1, Offset point2) {
   if (point1 == point2) {
     return null;
   }
-  final double x = point2.dx - point1.dx;
+  final double x = point2.dx - point1!.dx;
   final double y = point2.dy - point1.dy;
   return x.abs() > y.abs() ? Axis.horizontal : Axis.vertical;
 }
