@@ -8,19 +8,19 @@ class Disposable<T> {
   final T data;
 }
 
-T useDisposable<T>(T Function() disposableBuilder, [List<Object> keys]) {
+T useDisposable<T>(T Function() disposableBuilder, [List<Object> keys= const <Object>[]]) {
   final data = disposableBuilder();
   assert((data as dynamic).dispose is void Function());
 
   final d = Disposable(data, (data as dynamic).dispose as void Function());
-  return Hook.use(_DisposableHook(() => d, keys: keys));
+  return use(_DisposableHook(() => d, keys: keys));
 }
 
 class _DisposableHook<T> extends Hook<T> {
   final Disposable<T> Function() disposableBuilder;
 
   const _DisposableHook(this.disposableBuilder,
-      {List<Object> keys = const <dynamic>[]})
+      {List<Object> keys = const <Object>[]})
       : assert(disposableBuilder != null),
         assert(keys != null),
         super(keys: keys);
@@ -30,7 +30,7 @@ class _DisposableHook<T> extends Hook<T> {
 }
 
 class _DisposableHookState<T> extends HookState<T, _DisposableHook<T>> {
-  Disposable<T> disposable;
+  Disposable<T>? disposable;
 
   void createDisposable() {
     disposable = hook.disposableBuilder();
@@ -44,7 +44,7 @@ class _DisposableHookState<T> extends HookState<T, _DisposableHook<T>> {
 
   @override
   T build(BuildContext context) {
-    return disposable.data;
+    return disposable!.data;
   }
 
   @override
@@ -52,10 +52,10 @@ class _DisposableHookState<T> extends HookState<T, _DisposableHook<T>> {
     super.didUpdateHook(oldHook);
 
     final oldKeys = oldHook.keys;
-    final keys = hook.keys;
+    final keys = hook.keys!;
 
     final hasDifferentKeys =
-        Iterable<int>.generate(keys.length).any((i) => oldKeys[i] != keys[i]);
+        Iterable<int>.generate(keys.length).any((i) => oldKeys![i] != keys[i]);
     print(hasDifferentKeys);
     print(keys);
     print(oldKeys);
@@ -69,7 +69,7 @@ class _DisposableHookState<T> extends HookState<T, _DisposableHook<T>> {
   void dispose() {
     print("dispose!!");
     if (disposable != null) {
-      disposable.dispose();
+      disposable!.dispose();
     }
   }
 }
