@@ -9,7 +9,7 @@ part of 'input_layer.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
 
 mixin _$Input on _Input, Store {
-  Computed<Tensor> _$tensorComputed;
+  Computed<Tensor>? _$tensorComputed;
 
   @override
   Tensor get tensor => (_$tensorComputed ??=
