@@ -9,28 +9,13 @@ part of 'root_store.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
 
 mixin _$RootStore on _RootStore, Store {
-  Computed<String> _$generatedSourceCodeComputed;
+  Computed<String>? _$generatedSourceCodeComputed;
 
   @override
   String get generatedSourceCode => (_$generatedSourceCodeComputed ??=
           Computed<String>(() => super.generatedSourceCode,
               name: '_RootStore.generatedSourceCode'))
       .value;
-
-  final _$networksAtom = Atom(name: '_RootStore.networks');
-
-  @override
-  ObservableMap<String, NeuralNetwork> get networks {
-    _$networksAtom.reportRead();
-    return super.networks;
-  }
-
-  @override
-  set networks(ObservableMap<String, NeuralNetwork> value) {
-    _$networksAtom.reportWrite(value, super.networks, () {
-      super.networks = value;
-    });
-  }
 
   final _$selectedNetworkAtom = Atom(name: '_RootStore.selectedNetwork');
 
@@ -65,7 +50,6 @@ mixin _$RootStore on _RootStore, Store {
   @override
   String toString() {
     return '''
-networks: ${networks},
 selectedNetwork: ${selectedNetwork},
 language: ${language},
 generatedSourceCode: ${generatedSourceCode}
