@@ -3,7 +3,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 part 'async_result.freezed.dart';
 
 @freezed
-abstract class AsyncResult<S, E> implements _$AsyncResult<S, E> {
+class AsyncResult<S, E> with _$AsyncResult<S, E> {
   const AsyncResult._();
   const factory AsyncResult.idle() = _Idle<S, E>;
   const factory AsyncResult.loading() = _Loading<S, E>;
@@ -15,6 +15,6 @@ abstract class AsyncResult<S, E> implements _$AsyncResult<S, E> {
   bool get isSuccess => maybeWhen(orElse: () => false, success: (_) => true);
   bool get isError => maybeWhen(orElse: () => false, error: (_) => true);
 
-  S get valueOrNull => maybeWhen(orElse: () => null, success: (value) => value);
-  E get errorOrNull => maybeWhen(orElse: () => null, error: (error) => error);
+  S? get valueOrNull => maybeWhen(orElse: () => null, success: (value) => value);
+  E? get errorOrNull => maybeWhen(orElse: () => null, error: (error) => error);
 }
